Rename vc module to 1_1
