+++ crates/web5/src/apid/credentials/mod.rs
@@ -1,5 +1,5 @@
 pub mod presentation_definition;
-pub mod verifiable_credential_11;
+pub mod verifiable_credential_1_1;
 
 #[derive(thiserror::Error, Debug)]
 pub enum CredentialError {
